Begin hyperparameter exploration
Started recording the results of hyperparameter tuning to try to understand why the network isn't learning. Suspect
possibly the network's capacity is too small, so started training with a much  reduced training set, to try and see
if I can get it to overfit. If not, there may be a bug somewhere...